pyproject.toml
[tool.poetry]
name = "b4u-serving"
version = "0.1.0"
description = ""
authors = ["Andrey Korotkiy <Andreikorotkiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
pylint = "^2.17.4"
mypy = "^1.3.0"
pytest = "^7.3.1"
httpx = "^0.24.1"
boto3 = "^1.26.150"
python-dotenv = "^1.0.0"
scikit-learn = "^1.2.2"
redis = "^4.6.0"
pandas = "^2.0.3"
fastparquet = "^2023.7.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 89
disable = """
    too-many-return-statements,
    global-statement,
    raw-checker-failed,
    bad-inline-option,
    locally-disabled,
    file-ignored,
    suppressed-message,
    useless-suppression,
    deprecated-pragma,
    use-symbolic-message-instead,
    missing-module-docstring,
    invalid-name,
    missing-function-docstring,
    redefined-outer-name,
    import-error,
    missing-class-docstring,
    too-few-public-methods,
    attribute-defined-outside-init,
    too-many-locals,
    too-many-arguments,
    too-many-instance-attributes,
    unused-argument,
    no-member,
    arguments-differ,
    super-init-not-called
"""